General enhencement
- Split component structures
- Create Api class for requests
