+++ src/App/index_to_rename.tsx
@@ -1,0 +1,2 @@
+import App from "./App";
+export default App;
