--- src/services/Api.ts
@@ -1,5 +1,0 @@
-import axios from "axios";
-
-export default axios.create({
-  baseURL: process.env.REACT_APP_API_ENDPOINT,
-});
