--- src/store/store.ts
@@ -1,16 +1,0 @@
-import { configureStore, combineReducers } from "redux-starter-kit";
-
-import { userReducer } from "./user";
-import { roomReducer } from "./room";
-
-const rootReducer = combineReducers({
-  user: userReducer,
-  room: roomReducer,
-});
-
-const store = configureStore({
-  reducer: rootReducer,
-});
-
-export type StateStore = ReturnType<typeof rootReducer>;
-export default store;
