+++ src/store/utils/TypeOfConnect.ts
@@ -1,8 +1,10 @@
 import { InferableComponentEnhancerWithProps } from "react-redux";
 
-export type TypeOfConnect<T> = T extends InferableComponentEnhancerWithProps<
+type TypeOfConnect<T> = T extends InferableComponentEnhancerWithProps<
   infer Props,
   infer _
 >
   ? Props
   : never;
+
+export default TypeOfConnect;
