--- src/types/AuthInfo.ts
@@ -1,4 +1,0 @@
-export type AuthInfo = {
-  name: string;
-  token: string;
-};
